+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+- requires Terraform 1.x
+
+
+
 ## [0.10.0] - 2021-06-07
 
 ### Added