+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 - requires Terraform 1.x
+- assign_ipv6_address_on_creation silently ignored for non-IPv6 subnets, can be enabled in vpc/terraform.tfvars
 - bootstrap: use on-demand instead of provisioned capacity mode for DynamoDB table